update separacion
ahora está todo en archivos distintos
