+++ SSOO_tarea3/nbproject/project.xml
@@ -6,7 +6,7 @@
             <name>SSOO_tarea3</name>
             <c-extensions>c</c-extensions>
             <cpp-extensions/>
-            <header-extensions/>
+            <header-extensions>h</header-extensions>
             <sourceEncoding>UTF-8</sourceEncoding>
             <make-dep-projects/>
             <sourceRootList/>
